console: Refactor logic middleware to globalutil
